refactor(field_index): extract FieldIndexRead impl into separate file
